+++ CHANGELOG.md
@@ -9,6 +9,7 @@ See [Releases](https://gitlab.com/gitlab-org/step-runner/-/releases) for when ch
 - Steps can be loaded from the file system using an absolute path. See !215.
 - Steps can be loaded using an expression. See !215.
 - Fetching steps from an OCI registry uses a step to fetch. See !216.
+- Remove `Dir` from the proto `SpecDefinition` as it is a runtime property. See !212.
 
 ## v0.8.0